[Inductor][CPP] Fix Data Type issue of frexp
ghstack-source-id: 1342bef7c1e645c154f11703537187ae0b151398
Pull Request resolved: https://github.com/pytorch/pytorch/pull/143746

+++ torch/_inductor/codegen/cpp.py
@@ -1682,8 +1682,8 @@ class CppVecOverrides(CppOverrides):
         cdtype = DTYPE_TO_CPP[x.dtype]
         size = V.kernel.tail_size if V.kernel.tail_size else V.kernel.tiling_factor
         code = BracesBuffer()
-        exponent = V.kernel.cse.newvar()
-        mantissa = V.kernel.cse.newvar()
+        exponent = V.kernel.cse.newvar(dtype=torch.int32)
+        mantissa = V.kernel.cse.newvar(dtype=x.dtype)
         exponent.update_on_args("frexp", (x,), kwargs={})
         mantissa.update_on_args("frexp", (x,), kwargs={})
         n_vec = V.kernel._get_num_vectors(x.dtype)
